% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{analysis_duration}
\alias{analysis_duration}
\title{Duration Curve Analysis}
\usage{
analysis_duration(df, by_wycategory = FALSE, by_baseyear = FALSE, na.rm = TRUE)
}
\arguments{
\item{df}{A dataset.}

\item{by_baseyear}{If \code{TRUE}, compute metrics by base year.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing metrics.}
}
\value{
A dataframe with fields \code{"alt"}, (\code{"run"}), \code{"path"},
(\code{"period"}, \code{"wycategory"}, \code{"baseyear"}), \code{"metric"},
"probability"\verb{, and }"quantile"`.
}
\description{
Duration Curve Analysis
}
\keyword{internal}
