% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{analysis_frequency}
\alias{analysis_frequency}
\title{Frequency Curve Analysis}
\usage{
analysis_frequency(
  df,
  metrics = available_metrics(),
  by_wycategory = FALSE,
  by_baseyear = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{df}{A dataset.}

\item{metrics}{A vector of metrics to compute.}

\item{by_baseyear}{If \code{TRUE}, compute metrics by base year.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing metrics.}
}
\value{
A dataframe with fields \code{"alt"}, (\code{"run"}), \code{"path"},
(\code{"period"}, \code{"wycategory"}, \code{"baseyear"}), \code{"metric"},
\code{"probability"} and \code{"quantile"}.
}
\description{
Generate frequency analyses for each metric.
}
\keyword{internal}
