% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{analysis_metric_vs_forecast}
\alias{analysis_metric_vs_forecast}
\title{Metric vs Forecast Analysis}
\usage{
analysis_metric_vs_forecast(
  df.m,
  df.f,
  metrics,
  by_wycategory = FALSE,
  by_baseyear = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{df.m}{The dataset to calculate the metric on.}

\item{df.f}{The forecast dataset.}

\item{metrics}{A vector of metrics to compute.}

\item{by_baseyear}{If \code{TRUE}, compute metrics by base year.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing metrics.}
}
\value{
A dataframe with fields \code{"alt"}, (\code{"run"}),
\code{"path.m"}, \code{"path.f"}, (\code{"period.m"}, \code{"period.f"}),
\verb{"event}", \code{"date"}, (\code{"wateryear"}),
\code{"refdate"}, (\code{"wycategory"}, \code{"baseyear"}), \code{"metric"},
\code{"value.m"}, and \code{"value.f"}.
}
\description{
Calculate metrics of a dataset and compares against a forecast.
}
\keyword{internal}
