% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{analysis_metric_vs_metric}
\alias{analysis_metric_vs_metric}
\title{Metric vs Metric Analysis}
\usage{
analysis_metric_vs_metric(
  df.x,
  df.y,
  metrics.x,
  metrics.y,
  by_wycategory = FALSE,
  by_baseyear = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{df.x}{The dataset for the first axis.}

\item{df.y}{The dataset for the second axis.}

\item{metrics.x}{The metric for the first axis.}

\item{metrics.y}{The metric for the second axis.}

\item{by_baseyear}{If \code{TRUE}, compute metrics by base year.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing metrics.}
}
\value{
A dataframe with fields \code{"alt"}, (\code{"run"}), \code{"path.x"},
\code{"path.y"}, (\code{"period.x"}), (\code{"period.y"}), \code{"refdate"},
(\code{"wycategory"}, \code{"baseyear"}), \code{"metric.x"}, \code{"metric.y"},
\code{"value.x"}, and \code{"value.y"}.
}
\description{
Calculate metrics of two datasets and join by reference date.
}
\keyword{internal}
