% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{analysis_metrics}
\alias{analysis_metrics}
\title{Metric Over Time Analysis}
\usage{
analysis_metrics(
  df,
  metrics = available_metrics(),
  by_wycategory = FALSE,
  by_baseyear = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{df}{A dataset.}

\item{metrics}{A vector of metrics to compute.}

\item{by_baseyear}{If \code{TRUE}, compute metrics by base year.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing metrics.}

\item{by_wyclass}{If \code{TRUE}, compute metrics by water year category.}
}
\value{
A dataframe with fields \code{"alt"}, (\code{"run"}), \code{"path"},
(\code{"period"}), \code{"refdate"}, (\code{"wycategory"}, \code{"baseyear"}),
\code{"metric"}, and \code{"value"}.
}
\description{
Compute metrics over time for a dataset.
}
\keyword{internal}
