% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{analysis_trace}
\alias{analysis_trace}
\title{Trace Analysis}
\usage{
analysis_trace(
  df,
  sample = Inf,
  by_wycategory = FALSE,
  by_baseyear = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{df}{A dataset.}

\item{sample}{The maximum number of traces per group to return.}

\item{by_baseyear}{If \code{TRUE}, compute metrics by base year.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing metrics.}
}
\value{
A dataframe with fields \code{"alt"}, (\code{"run"}), \code{"path"},
(\code{"period"}), \code{"event"}, \code{"date"}, \code{"refdate"},
(\code{"wycategory"}, \code{"baseyear"}), and \code{"value"}.
}
\description{
Extract traces from a dataset.
}
\keyword{internal}
