% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damages.r
\name{annual_damages_simple}
\alias{annual_damages_simple}
\title{Expected Annual Damages (Simple)}
\usage{
annual_damages_simple(x)
}
\arguments{
\item{x}{A numeric vector of flows, in cfs.}
}
\value{
A vector of damages.
}
\description{
Compute Expected Annual Damages (EAD) as a simple linear function
of flow.
}
\details{
Damages are calculated using the formula
\itemize{
\item For \eqn{Q < 400, 000 \text{ cfs}}: \eqn{EAD = 0.01}
\item For \eqn{Q \geq 400, 000 \text{ cfs}}: \eqn{EAD = 10^{0.001mQ + b} - \alpha}
}

Where
\itemize{
\item \eqn{Q} is flow, in cfs
\item \eqn{m = 0.006127839}
\item \eqn{b = -3.301029996}
\item \eqn{\alpha = 0.1413}
}
}
\keyword{internal}
