% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{assert_match}
\alias{assert_match}
\title{Assert Unique Entries}
\usage{
assert_match(df.x, df.y, fields)
}
\arguments{
\item{df.x}{A dataset.}

\item{df.y}{A second dataset.}

\item{fields}{The fields to group by, i.e., output of \code{\link[=fields]{fields()}}.}
}
\value{
(Invisibly) TRUE if groups are unique, error otherwise.
}
\description{
Check that only one value for each field is present.
}
\keyword{internal}
