% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{assert_single}
\alias{assert_single}
\title{Assert Single Entries}
\usage{
assert_single(df, fields)
}
\arguments{
\item{df}{A dataset.}

\item{fields}{The fields to group by, i.e., output of \code{\link[=fields]{fields()}}.}
}
\value{
(Invisibly) TRUE if groups are unique, error otherwise.
}
\description{
Check that only one entry for each grouping is present.
}
\keyword{internal}
