% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{assert_unique}
\alias{assert_unique}
\title{Assert Unique Entries}
\usage{
assert_unique(df, fields)
}
\arguments{
\item{df}{A dataset.}

\item{fields}{The fields to group by, i.e., output of \code{\link[=fields]{fields()}}.}
}
\value{
(Invisibly) TRUE if groups are unique, error otherwise.
}
\description{
Check that only one value for each field is present.
}
\keyword{internal}
