% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{available_metrics}
\alias{available_metrics}
\title{Available Metrics}
\usage{
available_metrics()
}
\value{
A character vector of available metrics.
}
\description{
Metrics available for computation.
}
\details{
Descriptions of the various functions are included below:
\itemize{
\item \code{"p\% exceedance"}: return the \code{p}\% exceedance value based on the
CRT standard quantile calculation, i.e.,
\code{standard_quantile(x, probs = 1 - p)}. See \code{\link[=standard_quantile]{standard_quantile()}}
for more information. \verb{1\%}, \verb{10\%}, \verb{25\%}, \verb{50\%}, \verb{75\%}, \verb{90\%},
and \verb{99\%} exceedance probabilities are supported.
\item \code{"mean"}: Return the average value.
\item \code{"min"}: Return the minimum value.
\item \code{"max"}: Return the maximum value.
\item \code{"median"}: return the median value. Equivalent to \code{"50\% exceedance"}.
}
}
