% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{baseyear-aesthetics}
\alias{baseyear-aesthetics}
\alias{baseyear_lines}
\alias{baseyear_points}
\title{Base Year Geometry Aesthetics}
\usage{
baseyear_lines(baseyears = NULL)

baseyear_points(baseyears)
}
\value{
A list with elements "types" and "widths".
}
\description{
Aesthetics for base year geometries.
}
\keyword{internal}
