% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{build_aes_groups}
\alias{build_aes_groups}
\title{Build Aesthetic Groups}
\usage{
build_aes_groups(df, fields)
}
\arguments{
\item{df}{A dataset, i.e., output of \code{\link[=summary_analysis]{summary_analysis()}}.}

\item{fields}{A vector of fields to group.}
}
\value{
A character vector.
}
\description{
Create groups From a set of fields.
}
\keyword{internal}
