% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{check_groupings}
\alias{check_groupings}
\title{Check Groupings}
\usage{
check_groupings(metrics, by_wycategory, by_baseyear)
}
\arguments{
\item{metrics}{A vector of metrics to compute.}

\item{by_baseyear}{If \code{TRUE}, compute metrics by base year.}
}
\value{
\code{TRUE} (invisibily), or an error if grouping is invalid.
}
\description{
Check for valid grouping combination.
}
\keyword{internal}
