% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.r
\name{compute_duration}
\alias{compute_duration}
\title{Compute Duration}
\usage{
compute_duration(df, groups, probs = seq(0, 1, 0.001), na.rm = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{groups}{The grouping to use when computing durations.}

\item{probs}{The output (non-exceedance) probabilities.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing durations.}
}
\value{
A dataframe, with specified \code{groups} fields and additional
fields (non-exceedance) \code{"probability"} and \code{"quantile"}.
}
\description{
Compute durations from a dataset.
}
\keyword{internal}
