% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.r
\name{compute_frequency}
\alias{compute_frequency}
\title{Comput Frequency}
\usage{
compute_frequency(df, groups, probs = seq(0.001, 0.999, 0.001), na.rm = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{groups}{The grouping to use when computing frequencies.}

\item{probs}{The output (non-exceedance) probabilities.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing frequencies.}
}
\value{
A dataframe, with specified \code{groups} fields and additional
fields (non-exceedance) \code{"probability"} and \code{"quantile"}.
}
\description{
Compute frequencies from a dataset.
}
\keyword{internal}
