% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute Metrics}
\usage{
compute_metrics(df, metrics, groups, na.rm = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{metrics}{A vector of metrics to compute.}

\item{groups}{The grouping to use when computing metrics.}

\item{na.rm}{If \code{TRUE}, remove \code{NA} values before computing metrics.}
}
\value{
A dataframe, with field "date" is replaced with
"wydate".
}
\description{
Compute metrics for a dataset.
}
\keyword{internal}
