% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damages.r
\name{energy_price}
\alias{energy_price}
\title{Energy Price}
\usage{
energy_price(df, market)
}
\arguments{
\item{df}{A dataframe. Must contain fields fields \code{"month"}, \code{"event"},
and \code{"value"} (MWhr).}

\item{market}{A dataframe of monthly and event-based market prices.
Includes fields \code{"month"}, \code{"event"}, and \code{"price"} ($/MWhr).}
}
\description{
Calculate energy prices.
}
