% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{exceedance}
\alias{exceedance}
\title{Exceedance Probability Label}
\usage{
exceedance(percent = TRUE)
}
\arguments{
\item{percent}{If \code{TRUE}, format labels as percents.}
}
\value{
A labeling function.
}
\description{
Exceedance probability scale labels.
}
\keyword{internal}
