% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{extract_spec}
\alias{extract_spec}
\title{Extract Aesthetic Specification}
\usage{
extract_spec(group, spec, default)
}
\arguments{
\item{group}{The aesthetic group name.}

\item{spec}{A named vector of aesthetic specifications.}

\item{default}{The default aesthetic specification to use when no
match is found.}
}
\value{
An aesthetic specification
}
\description{
Extract aesthetic specification from a list based on string match.
}
\keyword{internal}
