% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{field-aesthetics}
\alias{field-aesthetics}
\alias{scale_x_refdate}
\alias{scale_x_duration}
\alias{scale_x_frequency}
\alias{scale_color_alt}
\alias{scale_x_value}
\alias{scale_y_value}
\alias{scale_line_group}
\alias{scale_point_group}
\title{Field Aesthetics}
\usage{
scale_x_refdate(xlabel)

scale_x_duration(xlabel)

scale_x_frequency(xlabel)

scale_color_alt(groups, alt_colors, guide)

scale_x_value(xlabel, xshift, xdivisor)

scale_y_value(ylabel, yshift, ydivisor)

scale_line_group(groups, guide, wycategories = NULL, baseyears = NULL)

scale_point_group(groups, guide, wycategories = NULL, baseyears = NULL)
}
\description{
Default aesthetics for fields.
}
\keyword{internal}
