% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{fields}
\alias{fields}
\title{Wat Plotter Fields}
\usage{
fields(...)
}
\arguments{
\item{...}{A vector of field codes, any of: "alt", "run", "path",
"event", "date", "value", "refdate", "wateryear", "wycategory",
"wypercentile", "baseyear", "baselabel", "metric", "probability",
"quantile".}
}
\value{
A vector of field names.
}
\description{
Standard Fields in WAT Plotter data frames. This is an extension of
\code{\link[watplotter.core:fields]{watplotter.core::fields()}} that includes additional fields produced
by analysis functions.
}
