% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{generate_colors}
\alias{generate_colors}
\title{Generate Colors}
\usage{
generate_colors(df, alt_colors)
}
\arguments{
\item{alt_colors}{A named list of colors to map to alternatives.
If missing, default colors will be generated with
\code{\link[scales:pal_brewer]{scales::brewer_pal()}} "Dark2" palette.}
}
\value{
A tibble of alternative (and possibly run) colors.
}
\description{
Generate colors for alternatives and runs. In cases where multiple
runs per alternative exist, a color gradient within each alternative
is generated.
}
\keyword{internal}
