% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{get_fields}
\alias{get_fields}
\alias{join_fields}
\title{Get Data Fields}
\usage{
get_fields(df, required, optional = NULL)

join_fields(df.x, df.y, required, optional)
}
\arguments{
\item{df}{A dataset.}

\item{required}{Required fields.}

\item{optional}{Optional fields.}

\item{df.x}{The dataset for the first axis.}

\item{df.y}{The dataset for the second axis.}
}
\value{
A character vector of field names.
}
\description{
Check for required fields and return optional fields if available.
}
\keyword{internal}
