% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{metric-aesthetics}
\alias{metric-aesthetics}
\alias{metric_lines}
\alias{metric_points}
\title{Metric Geometry Aesthetics}
\usage{
metric_lines()

metric_points()
}
\value{
A list with elements "types" and "widths".
}
\description{
Aesthetics for metric geometries.
}
\keyword{internal}
