% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metric_functions}
\alias{metric_functions}
\title{Metric Functions}
\usage{
metric_functions(metrics = available_metrics(), na.rm = TRUE)
}
\arguments{
\item{metrics}{A character vector of metric functions to return.
Default is to return all available metrics.}

\item{na.rm}{a logical indicating whether missing values should be
    removed.}
}
\value{
A named list of functions.
}
\description{
Get List of metric functions.
}
\keyword{internal}
