% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{normalize}
\alias{normalize}
\title{Normalize Vector}
\usage{
normalize(x, target = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{target}{The target value that the new vector will sum to.
Default is 1.}
}
\value{
The argument \code{x}, rescaled to sum to 1.
}
\description{
Normalize a vector so that it sums to a target value.
}
\keyword{internal}
