% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{ordered_cumsum}
\alias{ordered_cumsum}
\title{Ordered Cumulative Sum}
\usage{
ordered_cumsum(x, order_by = seq_along(x))
}
\arguments{
\item{x}{A numeric vector.}

\item{order_by}{An integer vector to order \code{x} by during the
cumulative sum. Default is to preserve order of \code{x}, resulting in
identical behavior to \code{\link[base:cumsum]{base::cumsum()}}. To order from smallest to
largest, use \code{order_by = order(x)}.}
}
\value{
The cumulative sum of `x with respect to the specified
ordering.
}
\description{
Perform a cumulative sum of a vector given a specified element order.
}
\keyword{internal}
