% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{path_to_label}
\alias{path_to_label}
\title{Path To Label Template}
\usage{
path_to_label(path)
}
\arguments{
\item{path}{A DSS Path.}
}
\value{
A string.
}
\description{
Create a standard axis label from a path.
}
\keyword{internal}
