% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_metric_vs_forecast}
\alias{plot_metric_vs_forecast}
\title{Metric vs Forecast Plot}
\usage{
plot_metric_vs_forecast(df, ..., alt_colors)
}
\arguments{
\item{df}{A dataset.}

\item{...}{
  Arguments passed on to \code{\link[=plot_overrides]{plot_overrides}}
  \describe{
    \item{\code{title}}{The plot title.}
    \item{\code{xlabel}}{The x-axis label.}
    \item{\code{xdivisor}}{A multiplier applied to the x-axis values.}
    \item{\code{xshift}}{an offset applied to the x-axis values.}
    \item{\code{ylabel}}{The y-axis label.}
    \item{\code{ydivisor}}{A multiplier applied to the y-axis values.}
    \item{\code{yshift}}{an offset applied to the y-axis values.}
  }}
}
\value{
A \code{ggplot} object.
}
\description{
Metric vs Forecast Plot
}
\keyword{internal}
