% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_metrics}
\alias{plot_metrics}
\title{Metrics Over Time Plot}
\usage{
plot_metrics(df, ..., alt_colors)
}
\arguments{
\item{df}{A dataset, i.e., output of \code{\link[=summary_analysis]{summary_analysis()}}.}

\item{...}{Additional arguments passed to \code{\link[=plot_overrides]{plot_overrides()}}.}

\item{alt_colors}{A named list of colors to map to alternatives.
If missing, default colors will be generated with
\code{\link[scales:pal_brewer]{scales::brewer_pal()}} "Dark2" palette.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a metrics over time analysis.
}
\keyword{internal}
