% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{plot_overrides}
\alias{plot_overrides}
\title{Plot Defaults and Overrides}
\usage{
plot_overrides(title, xlabel, xdivisor, xshift, ylabel, ydivisor, yshift)
}
\arguments{
\item{title}{The plot title.}

\item{xlabel}{The x-axis label.}

\item{xdivisor}{A multiplier applied to the x-axis values.}

\item{xshift}{an offset applied to the x-axis values.}

\item{ylabel}{The y-axis label.}

\item{ydivisor}{A multiplier applied to the y-axis values.}

\item{yshift}{an offset applied to the y-axis values.}
}
\value{
A list of overrides.
}
\description{
Aesthetic defaults and overrides.
}
\keyword{internal}
