% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{rescale_trans}
\alias{rescale_trans}
\title{Shift and Scale Tranformation}
\usage{
rescale_trans(shift = 0, divisor = 1)
}
\arguments{
\item{shift}{Numeric shift.}

\item{divisor}{Numeric divisor.}
}
\value{
A \code{scales} tranformation.
}
\description{
Basic WAT Plotter scale transform of the form \code{(x + shift) / divisor}.
}
\keyword{internal}
