% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{standard_quantile}
\alias{standard_quantile}
\title{CRT Standard Quantile}
\usage{
standard_quantile(x, probs, ...)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Standardized quantile calculation for CRT work. The primary change
from the default \code{\link[stats:quantile]{stats::quantile()}} is specification of \code{type = 5}
and removing NA values by default (\code{na.rm = TRUE}).
}
\keyword{internal}
