% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.r
\name{summary_analysis}
\alias{summary_analysis}
\title{Summary Analysis}
\usage{
summary_analysis(type = available_analyses(), ...)
}
\arguments{
\item{type}{The analysis type.}

\item{...}{
  Arguments passed on to \code{\link[=analysis_metrics]{analysis_metrics}}, \code{\link[=analysis_frequency]{analysis_frequency}}, \code{\link[=analysis_duration]{analysis_duration}}, \code{\link[=analysis_metric_vs_metric]{analysis_metric_vs_metric}}, \code{\link[=analysis_metric_vs_forecast]{analysis_metric_vs_forecast}}
  \describe{
    \item{\code{df}}{A dataset.}
    \item{\code{by_baseyear}}{If \code{TRUE}, compute metrics by base year.}
    \item{\code{metrics}}{A vector of metrics to compute.}
    \item{\code{na.rm}}{If \code{TRUE}, remove \code{NA} values before computing metrics.}
    \item{\code{df.x}}{The dataset for the first axis.}
    \item{\code{df.y}}{The dataset for the second axis.}
    \item{\code{metrics.x}}{The metric for the first axis.}
    \item{\code{metrics.y}}{The metric for the second axis.}
    \item{\code{df.m}}{The dataset to calculate the metric on.}
    \item{\code{df.f}}{The forecast dataset.}
  }}
}
\value{
A dataframe.
}
\description{
Generate a summary analysis. Provides a generic interface for
\code{\link[=analysis_metrics]{analysis_metrics()}}, \code{\link[=analysis_frequency]{analysis_frequency()}}, \code{\link[=analysis_duration]{analysis_duration()}},
\code{\link[=analysis_metric_vs_metric]{analysis_metric_vs_metric()}}, and \code{\link[=analysis_metric_vs_forecast]{analysis_metric_vs_forecast()}}.
}
