% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{summary_plot}
\alias{summary_plot}
\title{Summary Plot}
\usage{
summary_plot(type, df, ..., alt_colors)
}
\arguments{
\item{type}{The analysis type.}

\item{df}{A dataset, i.e., output of \code{\link[=summary_analysis]{summary_analysis()}}.}

\item{...}{
  Arguments passed on to \code{\link[=plot_metrics]{plot_metrics}}, \code{\link[=plot_frequency]{plot_frequency}}, \code{\link[=plot_duration]{plot_duration}}, \code{\link[=plot_metric_vs_metric]{plot_metric_vs_metric}}, \code{\link[=plot_metric_vs_forecast]{plot_metric_vs_forecast}}
  \describe{
    \item{\code{}}{}
  }}

\item{alt_colors}{A named list of colors to map to alternatives.
If missing, default colors will be generated with
\code{\link[scales:pal_brewer]{scales::brewer_pal()}} "Dark2" palette.}
}
\value{
A dataframe.
}
\description{
Visualize a summary analysis. Provides a generic interface for
\code{\link[=plot_metrics]{plot_metrics()}}, \code{\link[=plot_frequency]{plot_frequency()}}, \code{\link[=plot_duration]{plot_duration()}},
\code{\link[=plot_metric_vs_metric]{plot_metric_vs_metric()}}, and \code{\link[=plot_metric_vs_forecast]{plot_metric_vs_forecast()}}.
}
