% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{summary_render}
\alias{summary_render}
\title{Render Summary}
\usage{
summary_render(plot)
}
\arguments{
\item{plot}{A summary_analysis plot object, i.e., output of
\code{\link[=summary_plot]{summary_plot()}}.}
}
\value{
a \code{plotly} object, i.e., output of \code{\link[plotly:ggplotly]{plotly::ggplotly()}}.
}
\description{
Render a summary plot as an interactive HTML widget.
}
