% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{summary_save}
\alias{summary_save}
\title{Save Summary}
\usage{
summary_save(plot, filename, ...)
}
\arguments{
\item{plot}{A summary_analysis plot object, i.e., output of
\code{\link[=summary_plot]{summary_plot()}}.}

\item{filename}{The output filename, including extension. Supports
formats \code{"csv"}, \code{"png"}, \code{"html"}, and \code{"rds"}.}

\item{...}{For future expansion. Currently ignored.}
}
\value{
A file path.
}
\description{
Save a summary analysis.
}
