% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{theme_watplotter}
\alias{theme_watplotter}
\alias{guide_watplotter}
\title{WATPlotter Theme}
\usage{
theme_watplotter(add = list())

guide_watplotter(alts, groups, max_col = 3L)
}
\arguments{
\item{add}{Additional \code{\link[ggplot2:theme]{ggplot2::theme()}} elements.}
}
\description{
Default watplotter \code{ggplot} theme.
}
\keyword{internal}
