% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{watplotter.analysis-package}
\alias{watplotter.analysis}
\alias{watplotter.analysis-package}
\title{watplotter.analysis: WAT Plotter Data Analysis}
\description{
Generate standard analyses and plots from hydrologic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwbi-apps/watplotter.analysis}
  \item Report bugs at \url{https://github.com/cwbi-apps/watplotter.analysis/issues}
}

}
\author{
\strong{Maintainer}: Michael Koohafkan \email{Michael.C.Koohafkan@usace.army.mil}

Other contributors:
\itemize{
  \item Matthew Grzegorzewski \email{matthew.grzegorzewski@usace.army.mil} [contributor]
}

}
\keyword{internal}
