% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{weibull_position}
\alias{weibull_position}
\title{Weibull Plotting Position}
\usage{
weibull_position(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector of (non-exceedance) probabilities.
}
\description{
Calculate the Weibull plotting positions of a vector.
}
\details{
This implementation uses \code{\link[=rank]{rank()}} with the following
parameters:
\itemize{
\item Ties are ranked in order of occurrence in the data
(\code{ties.method = "first"})
\item \code{NA} values are put first.
}
}
\keyword{internal}
