% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{wycategory-aesthetics}
\alias{wycategory-aesthetics}
\alias{wycategory_lines}
\alias{wycategory_points}
\title{Water Year Category Geometry Aesthetics}
\usage{
wycategory_lines(wycategories = NULL)

wycategory_points(wycategories)
}
\value{
A list with elements "types" and "widths".
}
\description{
Aesthetics for water year category geometries.
}
\keyword{internal}
