% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.r
\name{zscore_trans}
\alias{zscore_trans}
\title{Z Score Tranformation}
\usage{
zscore_trans(reverse = TRUE)
}
\arguments{
\item{reverse}{If \code{TRUE}, also apply a reverse transformation.}
}
\value{
A \code{scales} tranformation.
}
\description{
Z Score scale transform.
}
\keyword{internal}
